test : test code refactor by change project structure
